+++ src/interfaces/cli/IFileData.d.ts
@@ -1,0 +1,4 @@
+export interface IFileData {
+    filePath: string;
+    content: string;
+}