+++ src/types/cli/TSourceCodeData.d.ts
@@ -1,0 +1,3 @@
+import { IFileData } from '../../interfaces/cli/IFileData';
+
+export type TSourceCodeData = string | IFileData[];