+++ CHANGELOG.md
@@ -4,6 +4,7 @@ v0.12.0
 ---
 * Added ability to disable and enable obfuscation for specific parts of the code by adding conditional comments. 
 * Added obfuscation of `es2015` class names.
+* CLI: added directory obfuscation.
 
 v0.11.2
 ---